+++ src/styles/buttonTheme.ts
@@ -1,0 +1,58 @@
+const button = {
+	BLUE: {
+		DISABLED: {
+			ICON: '#FFFFFF',
+			BG: '#C2D5FF',
+		},
+		DEFAULT: {
+			ICON: '#FFFFFF',
+			BG: '#3876F6',
+		},
+		HOVER: {
+			ICON: '#FFFFFF',
+			BG: '#245CCE',
+		},
+		PRESSED: {
+			ICON: '#90B2FF',
+			BG: '#0D47A1',
+		},
+	},
+	WHITE: {
+		DISABLED: {
+			ICON: '#C6C6D0',
+			BG: '#F4F6FA',
+		},
+		DEFAULT: {
+			ICON: '#212121',
+			BG: '#FFFFFF',
+		},
+		HOVER: {
+			ICON: '#212121',
+			BG: '#C6C6D0',
+		},
+		PRESSED: {
+			ICON: '#34343C',
+			BG: '#9090A0',
+		},
+	},
+	BLACK: {
+		DISABLED: {
+			ICON: '#9090A0',
+			BG: '#464656',
+		},
+		DEFAULT: {
+			ICON: '#FFFFFF',
+			BG: '#212121',
+		},
+		HOVER: {
+			ICON: '#FFFFFF',
+			BG: '#626273',
+		},
+		PRESSED: {
+			ICON: '#FFFFFF',
+			BG: '#34343C',
+		},
+	},
+};
+
+export default button;
